+++ CHANGES.rst
@@ -5,6 +5,7 @@
 
   - Protobuf with ``protoc`` [GH-1125]
   - systemd-analyze with ``systemd-analyze`` [GH-1135]
+  - Dockerfile with ``hadolint`` [GH-1194]
 
 - New features: